dev-requirements.txt
pytest>=3.6 # required for pytest-django
pytest-django
pytest-cov
# note: pytest-sugar is optional but pretty
pytest-sugar
django-debug-toolbar
sphinx
ansible
